+++ src/core/data/middlewares/index.ts
@@ -12,3 +12,4 @@
  */
 
 export * from 'core/data/middlewares/attach-mock';
+export * from 'core/data/middlewares/attach-status';